Feature/modify search date (#172)
* Modify default max month , division

* Add shinhancard default max month , division

* Refactoring code

* Run lint

* Modify test code

* Add research interval

+++ src/main/kotlin/com/rainist/collectcard/cardtransactions/dto/ListTransactionsRequest.kt
@@ -1,6 +1,6 @@
 package com.rainist.collectcard.cardtransactions.dto
 
-import com.rainist.collectcard.cardtransactions.CardTransactionServiceImpl.Companion.DEFAULT_MAX_MONTH
+import com.rainist.collectcard.common.service.CardOrganization.Companion.DEFAULT_MAX_MONTH
 import com.rainist.common.annotation.validation.StringDateFormat
 import com.rainist.common.util.DateTimeUtil
 import javax.validation.constraints.NotNull
@@ -19,7 +19,7 @@ data class ListTransactionsRequestDataBody(
 
     // 조회기간 시작일 YYYYMMDD default now() prev 12month
     @field:StringDateFormat(pattern = "yyyyMMdd")
-    var startAt: String = DateTimeUtil.kstNowLocalDate().minusMonths(DEFAULT_MAX_MONTH).let { DateTimeUtil.localDateToString(it, "yyyyMMdd") },
+    var startAt: String = DateTimeUtil.kstNowLocalDate().minusMonths(DEFAULT_MAX_MONTH.toLong()).let { DateTimeUtil.localDateToString(it, "yyyyMMdd") },
 
     // 조회기간 종료일 YYYYMMDD default now()
     @field:StringDateFormat(pattern = "yyyyMMdd")
